refactor: split control plane session mapping
